+++ NEWS.md
@@ -6,6 +6,9 @@
 - New functions: `coriolis()` and `coriolis.dy()`.
 - Faster `Anomaly()` (can't believe I did that).
 - `EOF()` principal componets are numeric instead of "PC1", etc...
+- For consistency with `stat_contour()` created `geom_contour_fill()` which 
+works just like `stat_contour_fill()`. 
+- Adde `stat_na()` for easy masking of `NA` values. 
 
 # meteoR 0.0.9100
 - Full (not perfect) documentation.